# docs/proxy-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "APInteract Proxy API",
    "summary": "Public contract for the APInteract proxy component.",
    "description": "This specification is the canonical source of truth for the APInteract proxy component API used by the backend.",
    "version": "0.1.1",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
...
  "servers": [
    {
      "url": "https://{proxyHost}",
      "description": "Standalone proxy deployment. Remote proxy communication must use TLS.",
      "variables": {
        "proxyHost": {
          "default": "proxy.local",
          "description": "Operator-configured proxy hostname."
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/capabilities": {
      "get": {
        "tags": ["Proxy service"],
        "operationId": "getProxyCapabilities",
        "summary": "Get proxy capabilities and principal limits",
        "description": "Returns protocol support and effective limits for the authenticated backend principal. The backend must verify API and frame compatibility before creating executions.",
        "responses": {
          "200": {
            "description": "Capabilities and effective limits.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/executions": {
      "post": {
        "tags": ["Proxy executions"],
        "operationId": "createProxyExecution",
        "summary": "Create an HTTP execution session",
        "description": "Creates a transient execution owned by the authenticated backend principal. Ownership is derived exclusively from the bearer token and cannot be supplied or changed by the client. Idempotency keys are scoped to the principal and retained until the execution is released or expires.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExecutionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Execution session created, or the original creation result replayed for the same principal, idempotency key, and request descriptor.",
            "headers": {
              "Location": {
                "description": "Relative or absolute URL of the created execution.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              },
              "Idempotency-Replayed": {
                "description": "True when this response replays an execution created by an earlier request with the same idempotency key.",
                "schema": {
                  "type": "boolean"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionSession"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/IdempotencyConflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/executions/{executionId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ExecutionId"
        }
      ],
      "get": {
        "tags": ["Proxy executions"],
        "operationId": "getProxyExecution",
        "summary": "Get execution state",
        "description": "Returns current control-plane state for an execution owned by the authenticated principal. A missing execution and an execution owned by another principal both return 404.",
        "responses": {
          "200": {
            "description": "Current execution state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionSession"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/ExecutionNotFound"
          }
        }
      },
      "delete": {
        "tags": ["Proxy executions"],
        "operationId": "releaseProxyExecution",
        "summary": "Release a terminal execution",
        "description": "Acknowledges that the backend has persisted or discarded the result and releases the execution, its response cache, and its idempotency-key mapping. Active executions cannot be released and must be cancelled first.",
        "responses": {
          "204": {
            "description": "The terminal execution and its transient data were released."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/ExecutionNotFound"
          },
          "409": {
            "$ref": "#/components/responses/ExecutionStateConflict"
          }
        }
      }
    },
    "/executions/{executionId}/request-body": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ExecutionId"
        }
      ],
      "put": {
        "tags": ["Proxy executions"],
        "operationId": "uploadProxyExecutionRequestBody",
        "summary": "Stream the target request body",
        "description": "Streams the target request payload as unmodified bytes. This operation is valid only for an execution created with body mode `stream` and while that execution is accepting an upload. The authenticated principal must own the execution. The protocol content type does not replace or imply the target request's Content-Type header. Request upload resume is not supported in framing version 1; an interrupted upload fails the execution.",
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              },
              "x-apinteract-streaming": {
                "direction": "backend-to-proxy",
                "preserveBytes": true,
                "supportsBackpressure": true,
                "transferLengthMayBeUnknown": true
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The complete request body was accepted."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/ExecutionNotFound"
          },
          "409": {
            "$ref": "#/components/responses/ExecutionStateConflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        }
      }
    },
    "/executions/{executionId}/response": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ExecutionId"
        }
      ],
      "get": {
        "tags": ["Proxy executions"],
        "operationId": "streamProxyExecutionResponse",
        "summary": "Stream the target response",
        "description": "Opens the response data plane for an execution owned by the authenticated principal. Only one response reader may be active for an execution. The proxy returns a framed binary stream and may send heartbeat frames while waiting for the target. A RESPONSE_HEAD frame announces the target response before any BODY frame. The outer HTTP status describes this proxy operation; every valid target HTTP status, including 4xx and 5xx, is carried in RESPONSE_HEAD and completes normally unless a separate network or protocol failure occurs. Cached frames remain available until the terminal execution is released or expires.",
        "parameters": [
          {
            "$ref": "#/components/parameters/AfterSequence"
          }
        ],
        "responses": {
          "200": {
            "description": "Framed target response stream. A terminal COMPLETE or ERROR frame is required; connection closure alone does not indicate successful completion.",
            "content": {
              "application/vnd.apinteract.proxy-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyResponseStream"
                },
                "x-apinteract-framing": {
                  "version": 1,
                  "byteOrder": "big-endian",
                  "jsonEncoding": "UTF-8 without BOM",
                  "frameHeaderBytes": 16,
                  "maxFramePayloadBytes": 1048576,
                  "protocolRules": [
                    "Every frame, including HEARTBEAT, consumes one sequence number.",
                    "Sequence starts at zero, increments by one, and must not wrap.",
                    "Flags and reserved fields must be zero.",
                    "Unknown frame types, nonzero flags, invalid sequence values, oversized payloads, invalid JSON, and schema-invalid JSON are fatal protocol errors.",
                    "A receiver must reject an oversized payload length before allocating payload storage.",
                    "RESPONSE_HEAD must precede BODY and may occur only once.",
                    "TRAILERS may occur only after all BODY frames and before COMPLETE.",
                    "Exactly one COMPLETE or ERROR terminal frame is required.",
                    "No frames may follow a terminal frame, and the stream must end after it.",
                    "End of stream before a terminal frame is proxy_stream_incomplete.",
                    "Replayed streams preserve original frame sequence numbers."
                  ],
                  "headerFields": [
                    {
                      "name": "type",
                      "offset": 0,
                      "size": 1,
                      "format": "uint8"
                    },
                    {
                      "name": "flags",
                      "offset": 1,
                      "size": 1,
                      "format": "uint8",
                      "requiredValue": 0
                    },
                    {
                      "name": "reserved",
                      "offset": 2,
                      "size": 2,
                      "format": "uint16",
                      "requiredValue": 0
                    },
                    {
                      "name": "sequence",
                      "offset": 4,
                      "size": 4,
                      "format": "uint32",
                      "description": "Starts at zero and increments by one for every frame."
                    },
                    {
                      "name": "payloadLength",
                      "offset": 8,
                      "size": 8,
                      "format": "uint64",
                      "description": "Number of payload bytes immediately following the header."
                    }
                  ],
                  "frameTypes": [
                    {
                      "value": 1,
                      "name": "RESPONSE_HEAD",
                      "payloadMediaType": "application/json",
                      "payloadSchema": {
                        "$ref": "#/components/schemas/ResponseHead"
                      },
                      "cardinality": "exactly once before the first BODY frame when a target response is received"
                    },
                    {
                      "value": 2,
                      "name": "BODY",
                      "payloadMediaType": "application/octet-stream",
                      "cardinality": "zero or more after RESPONSE_HEAD and before a terminal frame"
                    },
                    {
                      "value": 3,
                      "name": "TRAILERS",
                      "payloadMediaType": "application/json",
                      "payloadSchema": {
                        "$ref": "#/components/schemas/ResponseTrailers"
                      },
                      "cardinality": "zero or one after all BODY frames and before COMPLETE"
                    },
                    {
                      "value": 4,
                      "name": "COMPLETE",
                      "payloadMediaType": "application/json",
                      "payloadSchema": {
                        "$ref": "#/components/schemas/ResponseComplete"
                      },
                      "cardinality": "exactly one terminal frame on success"
                    },
                    {
                      "value": 5,
                      "name": "ERROR",
                      "payloadMediaType": "application/json",
                      "payloadSchema": {
                        "$ref": "#/components/schemas/ExecutionStreamError"
                      },
                      "cardinality": "exactly one terminal frame on failure"
                    },
                    {
                      "value": 6,
                      "name": "HEARTBEAT",
                      "payloadMediaType": "application/octet-stream",
                      "payloadLength": 0,
                      "cardinality": "zero or more before a terminal frame"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/ExecutionNotFound"
          },
          "409": {
            "$ref": "#/components/responses/ExecutionStateConflict"
          },
          "410": {
            "$ref": "#/components/responses/ExecutionExpired"
          },
          "416": {
            "$ref": "#/components/responses/SequenceNotSatisfiable"
          }
        }
      }
    },
    "/executions/{executionId}/cancel": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ExecutionId"
        }
      ],
      "post": {
        "tags": ["Proxy executions"],
        "operationId": "cancelProxyExecution",
        "summary": "Cancel an execution",
        "description": "Requests best-effort cancellation of request upload, target communication, response streaming, and transient caching for an execution owned by the authenticated principal. Cancellation cannot guarantee that the target did not already process some or all of the request. Repeated cancellation requests are idempotent and return the current session state.",
        "responses": {
          "200": {
            "description": "The execution was already terminal and its state was unchanged.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionSession"
                }
              }
            }
          },
          "202": {
            "description": "Cancellation was accepted or had already been requested.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionSession"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/ExecutionNotFound"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token mapped by proxy configuration to a stable backend principal. Tokens must be sent only over TLS. Execution ownership is associated with the configured principal identifier, not with the literal token value or an APInteract application user."
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Opaque caller-generated key scoped to the authenticated principal. Concurrent or repeated creation requests with the same key and identical descriptor resolve to one execution. Reuse with different content returns 409.",
        "schema": {
          "type": "string",
          "minLength": 16,
          "maxLength": 128,
          "pattern": "^[!-~]+$"
        }
      },
      "ExecutionId": {
        "name": "executionId",
        "in": "path",
        "required": true,
        "description": "Opaque identifier of a transient proxy execution.",
        "schema": {
          "$ref": "#/components/schemas/ExecutionId"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "The request does not satisfy the proxy protocol.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "The bearer token is missing, invalid, or expired.",
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string"
            },
            "example": "Bearer"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "IdempotencyConflict": {
        "description": "The idempotency key was previously used by this principal with a different execution descriptor.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "PayloadTooLarge": {
        "description": "The request exceeds a configured metadata, body, or cache limit.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "The authenticated principal has exceeded a configured execution or transfer limit.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Seconds before another attempt should be made."
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "The proxy cannot accept another execution.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Seconds before another attempt should be made."
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Capabilities": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "apiVersion",
          "responseFrameVersions",
          "outboundHttpVersions",
          "features",
          "limits"
        ],
        "properties": {
          "apiVersion": {
            "type": "string",
            "const": "0.1.1"
          },
          "responseFrameVersions": {
            "type": "array",
            "minItems": 1,
            "uniqueItems": true,
            "items": {
              "type": "integer",
              "enum": [1]
            }
          },
          "outboundHttpVersions": {
            "type": "array",
            "minItems": 1,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "enum": ["HTTP/1.1"]
            }
          },
          "features": {
            "$ref": "#/components/schemas/CapabilityFeatures"
          },
          "limits": {
            "$ref": "#/components/schemas/CapabilityLimits"
          }
        }
      },
...
      "CreateExecutionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["request"],
        "properties": {
          "request": {
            "$ref": "#/components/schemas/TargetRequest"
          }
        }
      },
...
      "ExecutionSession": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "executionId",
          "state",
          "requestBodyState",
          "responseState",
          "createdAt",
          "expiresAt",
          "error"
        ],
        "properties": {
          "executionId": {
            "$ref": "#/components/schemas/ExecutionId"
          },
          "state": {
            "$ref": "#/components/schemas/ExecutionState"
          },
          "requestBodyState": {
            "$ref": "#/components/schemas/RequestBodyState"
          },
          "responseState": {
            "$ref": "#/components/schemas/ResponseState"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": ["string", "null"],
            "format": "date-time",
            "description": "Null while the execution is active. After a terminal state, the time at which the proxy may delete transient state, cached frames, and the idempotency-key mapping."
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionStreamError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Safe terminal failure details. Null unless state is failed."
          }
        },
        "description": "Transient execution state. Principal identity is intentionally omitted because ownership is derived from authentication and is not client-controlled.",
        "x-apinteract-state-rules": [
          "A bodyless execution becomes active when it is created.",
          "A streaming-body execution becomes active when request-body upload begins.",
          "Request upload and target response streaming may overlap.",
          "Completed, failed, cancelled, and expired are terminal states.",
          "Only terminal executions may be released.",
          "Cancellation is best effort and cannot retract bytes already sent to the target."
        ]
      },
...
    }
  }
}